Use RefCell on ComponentMaps
A lot of our operations on components will need access to multiple
components at once (with a soup of mutability and immutability), so we
want to punt the borrow checking to runtime.

Time will tell if the runtime overhead of RefCells will be too much, in
which case, we might need to switch to an index-based approach.