Refactor ECS to map from entity to component
Before, there was a separate GenerationalIndexArray for every component,
and you indexed into it with the entity.  This made it difficult to
filter by component on systems, because when you wanted to check if an
entity had a component, you couldn't use the type ID to specify the
generic type parameter on the GenerationalIndexArray, so it wouldn't
compile.